migrations checked

+++ db/migrate/002_extend_sites.rb
@@ -1,13 +1,17 @@
 class ExtendSites < ActiveRecord::Migration
   def self.up
-    add_column :sites, :reader_layout_id, :integer
-    add_column :sites, :mail_from_name, :string
-    add_column :sites, :mail_from_address, :string
+    if defined? Site
+      add_column :sites, :reader_layout_id, :integer
+      add_column :sites, :mail_from_name, :string
+      add_column :sites, :mail_from_address, :string
+    end
   end
 
   def self.down
-    remove_column :sites, :reader_layout_id
-    remove_column :sites, :mail_from_name
-    remove_column :sites, :mail_from_address
+    if defined? Site
+      remove_column :sites, :reader_layout_id
+      remove_column :sites, :mail_from_name
+      remove_column :sites, :mail_from_address
+    end
   end
 end
